most of saga working

+++ loan-demo-model/src/main/java/com/acme/saga/sagamodel/service/LoanServiceImpl.java
@@ -2,20 +2,37 @@ package com.acme.saga.sagamodel.service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpHeaders;
+import org.springframework.dao.EmptyResultDataAccessException;
 import com.acme.saga.sagamodel.helpers.Utils;
 import com.acme.saga.sagamodel.model.ApplicantRepository;
 import com.acme.saga.sagamodel.model.Loan;
 import com.acme.saga.sagamodel.model.LoanRepository;
+import com.acme.saga.sagamodel.dto.DeleteLoanResponseDTO;
+import lombok.extern.slf4j.Slf4j;
 
-
+@Slf4j
 @Service
 public class LoanServiceImpl implements LoanService {
 
     @Autowired LoanRepository loanRepository;    
+
+    @Override
+    public Loan findLoanById( Integer Id ) {
+        Optional<Loan> rLoan = loanRepository.findById( Id );
+
+        if(rLoan.isPresent())
+            return rLoan.get();
+        else
+            return null;
+    }
+
 
     @Override
     public Loan saveLoan( Loan loan ) {
@@ -68,8 +85,44 @@ public class LoanServiceImpl implements LoanService {
 
 
     @Override
-    public void deleteLoanById(Integer Id) {
-        loanRepository.deleteById( Id );
+    public ResponseEntity<DeleteLoanResponseDTO> deleteLoanById(Integer Id) {
+        // setup response entity for return 
+        HttpHeaders responseHeaders = new HttpHeaders();
+        
+        responseHeaders.add("acme-loan-id", Id.toString());
+
+        DeleteLoanResponseDTO rDTO = DeleteLoanResponseDTO.builder()
+                                                          .loanId(Id)
+                                                          .build();
+        ResponseEntity<DeleteLoanResponseDTO> rEntity = null;
+
+        Optional<Loan> theLoan = loanRepository.findById( Id );
+
+        if(!theLoan.isPresent()) {
+            log.error("Trying to delete Loan with ID: " + Id.toString() + ", which does not exist...");
+            rDTO.setComment("Unable to find Loan with ID: " + Id.toString());
+            rDTO.setStatus(HttpStatus.NOT_FOUND);
+            responseHeaders.add("acme-loan-delete", "notfound");
+            rEntity = new ResponseEntity<>(rDTO, responseHeaders, HttpStatus.NOT_FOUND);
+            return rEntity;
+        }
+
+        try {
+            loanRepository.deleteById( Id );
+        } catch (EmptyResultDataAccessException e) {
+            rDTO.setComment("Unable to find Loan with ID: " + Id.toString());
+            rDTO.setStatus(HttpStatus.NOT_FOUND);
+            responseHeaders.add("acme-loan-delete", "notfound");
+            rEntity = new ResponseEntity<>(rDTO, responseHeaders, HttpStatus.NOT_FOUND);
+            return rEntity;
+        }
+
+        responseHeaders.add("acme-loan-delete", "success");
+        rDTO.setComment("Loan with ID: " + Id.toString() + " deleted...");
+        rDTO.setStatus(HttpStatus.OK);
+        rEntity = new ResponseEntity<>(rDTO, responseHeaders, HttpStatus.OK);
+
+        return rEntity;
     }
 
 }
